% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-formatting.R
\name{apply_zzo2f_format}
\alias{apply_zzo2f_format}
\title{Apply a single zzo2f_format spec to a tinytable object}
\usage{
apply_zzo2f_format(tbl, spec, data)
}
\arguments{
\item{tbl}{A \code{tinytable} object.}

\item{spec}{A \code{zzo2f_format} spec.}

\item{data}{The data frame backing \code{tbl}.}
}
\value{
Modified \code{tinytable} object.
}
\description{
Apply a single zzo2f_format spec to a tinytable object
}
\keyword{internal}
