% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-formatting.R
\name{apply_zzo2f_formats}
\alias{apply_zzo2f_formats}
\title{Apply a list of zzo2f_format specs to a tinytable object}
\usage{
apply_zzo2f_formats(tbl, formats, data)
}
\arguments{
\item{tbl}{A \code{tinytable} object.}

\item{formats}{A single \code{zzo2f_format}, a list of them, or
NULL.}

\item{data}{The data frame backing \code{tbl}.}
}
\value{
Modified \code{tinytable} object.
}
\description{
Apply a list of zzo2f_format specs to a tinytable object
}
\keyword{internal}
