% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-preamble.R
\name{fmt_margin}
\alias{fmt_margin}
\title{Format a margin argument as a Typst expression}
\usage{
fmt_margin(margin)
}
\arguments{
\item{margin}{Character or named list.}
}
\value{
Character scalar.
}
\description{
Format a margin argument as a Typst expression
}
\keyword{internal}
