% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{o2f.anova}
\alias{o2f.anova}
\alias{o2f.aov}
\title{o2f method for ANOVA objects}
\usage{
\method{o2f}{anova}(x, digits = 3, ...)

\method{o2f}{aov}(x, digits = 3, ...)
}
\arguments{
\item{x}{An anova or aov object.}

\item{digits}{Number of decimal places.}

\item{...}{Additional arguments passed to o2f.default.}
}
\value{
Invisibly returns the path to the generated PDF.
}
\description{
o2f method for ANOVA objects

o2f method for aov objects
}
