% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{o2f.htest}
\alias{o2f.htest}
\title{o2f method for hypothesis tests}
\usage{
\method{o2f}{htest}(x, digits = 3, ...)
}
\arguments{
\item{x}{An htest object (from t.test, chisq.test, etc.).}

\item{digits}{Number of decimal places.}

\item{...}{Additional arguments passed to o2f.default.}
}
\value{
Invisibly returns the path to the generated PDF.
}
\description{
o2f method for hypothesis tests
}
\examples{
\dontrun{
test <- t.test(mtcars$mpg, mu = 20)
o2f(test, filename = "ttest_result")
}

}
