% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{o2f.table}
\alias{o2f.table}
\title{o2f method for table}
\usage{
\method{o2f}{table}(x, ...)
}
\arguments{
\item{x}{A table object to convert to a LaTeX table.}

\item{...}{Additional arguments passed to o2f.default.}
}
\value{
Invisibly returns the path to the generated PDF.
}
\description{
o2f method for table
}
