% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{o2f_bold_col}
\alias{o2f_bold_col}
\title{Bold specific columns}
\usage{
o2f_bold_col(cols)
}
\arguments{
\item{cols}{Integer vector or character vector of column indices or names.}
}
\value{
A o2f_format object with bold=TRUE for specified columns.
}
\description{
Convenience function to bold entire columns.
}
\examples{
\dontrun{
o2f_bold_col(c("estimate", "p_value"))
o2f_bold_col(1:2)
}

}
