% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{o2f_cache_dir}
\alias{o2f_cache_dir}
\title{Get the cache directory}
\usage{
o2f_cache_dir(create = TRUE)
}
\arguments{
\item{create}{Logical. Create the directory if it doesn't exist.}
}
\value{
Path to the cache directory.
}
\description{
Returns the path to the o2f cache directory. Creates it if
it doesn't exist.
}
