% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{o2f_cache_info}
\alias{o2f_cache_info}
\title{Get cache statistics}
\usage{
o2f_cache_info()
}
\value{
A list with cache statistics.
}
\description{
Report on the current state of the o2f cache.
}
\examples{
\dontrun{
o2f_cache_info()
}

}
