% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{o2f_cache_set_dir}
\alias{o2f_cache_set_dir}
\title{Set the cache directory}
\usage{
o2f_cache_set_dir(path)
}
\arguments{
\item{path}{Path to use for caching. Set to NULL to use the default
(temp directory).}
}
\value{
Invisibly returns the previous cache directory.
}
\description{
Set the cache directory
}
