% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-features.R
\name{o2f_decimal}
\alias{o2f_decimal}
\title{Create decimal-aligned column specification (convenience wrapper)}
\usage{
o2f_decimal(integers = 3, decimals = 2)
}
\arguments{
\item{integers}{Maximum integer digits before decimal point.}

\item{decimals}{Maximum decimal digits after decimal point.}
}
\value{
A o2f_siunitx object.
}
\description{
Convenience function to create decimal-aligned columns.
Returns alignment string and updates extra_packages.
}
\examples{
\dontrun{
# Align numbers up to 99.999
o2f(df, align = c("l", o2f_decimal(2, 3), o2f_decimal(2, 3)))
}

}
