% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{o2f_italic_col}
\alias{o2f_italic_col}
\title{Italicize specific columns}
\usage{
o2f_italic_col(cols)
}
\arguments{
\item{cols}{Integer vector or character vector of column indices or names.}
}
\value{
A o2f_format object with italic=TRUE for specified columns.
}
\description{
Convenience function to italicize entire columns.
}
\examples{
\dontrun{
o2f_italic_col("variable")
}

}
