% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex-include.R
\name{o2f_margin_packages}
\alias{o2f_margin_packages}
\title{Get required LaTeX packages for margin methods}
\usage{
o2f_margin_packages(
  method = c("sidenotes", "marginpar", "tufte", "marginnote")
)
}
\arguments{
\item{method}{The margin method: "sidenotes", "marginpar", "tufte",
or "marginnote".}
}
\value{
Character string with the required \\usepackage command(s),
or NULL if no packages needed.
}
\description{
Returns the LaTeX package requirements for each margin
placement method.
}
\examples{
o2f_margin_packages("sidenotes")

}
