% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-features.R
\name{o2f_mark}
\alias{o2f_mark}
\title{Create footnote marker for table cells}
\usage{
o2f_mark(text, mark, type = c("symbol", "number", "alphabet"))
}
\arguments{
\item{text}{The cell text to add a marker to.}

\item{mark}{The footnote marker (number, letter, or symbol index).}

\item{type}{Marker type: "number", "alphabet", or "symbol".}
}
\value{
Character string with LaTeX footnote marker.
}
\description{
Insert a footnote marker into a cell value. Use this to mark
cells that correspond to footnotes.
}
\examples{
\dontrun{
df$value[1] <- o2f_mark("23.5", 1, "symbol")
# Produces: "23.5*" (with proper LaTeX superscript)
}

}
