% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-formats.R
\name{o2f_output_formats}
\alias{o2f_output_formats}
\title{List available output formats}
\usage{
o2f_output_formats()
}
\value{
Named logical vector indicating availability of each format.
}
\description{
Check which output formats are available based on installed
system tools.
}
\examples{
\dontrun{
o2f_output_formats()
}

}
