% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{o2f_theme_clear}
\alias{o2f_theme_clear}
\title{Clear all custom themes}
\usage{
o2f_theme_clear()
}
\value{
Invisibly returns the number of themes removed.
}
\description{
Remove all registered custom themes, restoring only built-in
themes.
}
