% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{o2f_theme_get}
\alias{o2f_theme_get}
\title{Get the current global o2f theme}
\usage{
o2f_theme_get()
}
\value{
The current o2f_theme object, or NULL if no theme is set.
}
\description{
Retrieve the currently active global theme.
}
