% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{o2f_theme_nature}
\alias{o2f_theme_nature}
\title{Nature theme}
\usage{
o2f_theme_nature()
}
\value{
A o2f_theme object.
}
\description{
Theme following Nature journal table styling. Uses Helvetica
font, light gray row shading, and compact spacing.
}
