% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-methods.R
\name{o2f_tidy}
\alias{o2f_tidy}
\title{Convert any broom-supported object to a LaTeX table}
\usage{
o2f_tidy(x, tidy_args = list(), glance = FALSE, digits = 3, ...)
}
\arguments{
\item{x}{An object supported by broom::tidy().}

\item{tidy_args}{A list of arguments to pass to broom::tidy().}

\item{glance}{Logical. If TRUE, append model-level statistics from
broom::glance() as footnote or additional rows.}

\item{digits}{Number of decimal places for rounding.}

\item{...}{Additional arguments passed to o2f().}
}
\value{
Invisibly returns the path to the generated PDF.
}
\description{
A general-purpose function that uses broom::tidy() to convert
any broom-supported object to a tidy data frame, then generates a LaTeX

table via o2f().
}
\examples{
\dontrun{
library(survival)
cox_model <- coxph(Surv(time, status) ~ age + sex, data = lung)
o2f_tidy(cox_model, filename = "cox_results")
}

}
