% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-postprocess.R
\name{pp_caption_above}
\alias{pp_caption_above}
\title{Move caption above table}
\usage{
pp_caption_above(lines)
}
\arguments{
\item{lines}{Character vector.}
}
\value{
Modified character vector.
}
\description{
Move caption above table
}
\keyword{internal}
