% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-postprocess.R
\name{pp_compact_footnotes}
\alias{pp_compact_footnotes}
\title{Style footnotes to be compact}
\usage{
pp_compact_footnotes(lines)
}
\arguments{
\item{lines}{Character vector.}
}
\value{
Modified character vector.
}
\description{
Style footnotes to be compact
}
\keyword{internal}
