% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-postprocess.R
\name{pp_font_family}
\alias{pp_font_family}
\title{Insert font family directive}
\usage{
pp_font_family(lines, font_family)
}
\arguments{
\item{lines}{Character vector.}

\item{font_family}{Character or NULL.}
}
\value{
Modified character vector.
}
\description{
Insert font family directive
}
\keyword{internal}
