% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-postprocess.R
\name{pp_page_size}
\alias{pp_page_size}
\title{Prepend page size directive}
\usage{
pp_page_size(lines)
}
\arguments{
\item{lines}{Character vector.}
}
\value{
Modified character vector.
}
\description{
Prepend page size directive
}
\keyword{internal}
