% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-structure.R
\name{print.zzo2f_collapse}
\alias{print.zzo2f_collapse}
\title{Print method for zzo2f_collapse}
\usage{
\method{print}{zzo2f_collapse}(x, ...)
}
\arguments{
\item{x}{A \code{zzo2f_collapse} object.}

\item{...}{Ignored.}
}
\value{
Invisibly \code{x}.
}
\description{
Print method for zzo2f_collapse
}
