% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-formatting.R
\name{print.zzo2f_format}
\alias{print.zzo2f_format}
\title{Print method for zzo2f_format}
\usage{
\method{print}{zzo2f_format}(x, ...)
}
\arguments{
\item{x}{A \code{zzo2f_format} object.}

\item{...}{Ignored.}
}
\value{
Invisibly \code{x}.
}
\description{
Print method for zzo2f_format
}
