% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-siunitx.R
\name{print.zzo2f_siunitx}
\alias{print.zzo2f_siunitx}
\title{Print method for zzo2f_siunitx}
\usage{
\method{print}{zzo2f_siunitx}(x, ...)
}
\arguments{
\item{x}{A \code{zzo2f_siunitx} object.}

\item{...}{Ignored.}
}
\value{
Invisibly \code{x}.
}
\description{
Print method for zzo2f_siunitx
}
