% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-engine.R
\name{register_o2f_engine}
\alias{register_o2f_engine}
\title{Register the o2f knitr engine}
\usage{
register_o2f_engine()
}
\value{
Invisible NULL.
}
\description{
Register the o2f engine with knitr. This is called
automatically when the package is loaded.
}
