% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{resolve_theme}
\alias{resolve_theme}
\title{Resolve theme from argument or global setting}
\usage{
resolve_theme(theme)
}
\arguments{
\item{theme}{Theme argument from o2f() call (can be NULL, character, or
o2f_theme object).}
}
\value{
A o2f_theme object or NULL.
}
\description{
Resolve theme from argument or global setting
}
\keyword{internal}
