% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-formatting.R
\name{style_tt_spec}
\alias{style_tt_spec}
\title{Dispatch a single \code{style_tt} call for a format spec}
\usage{
style_tt_spec(tbl, i, j, spec)
}
\description{
Dispatch a single \code{style_tt} call for a format spec
}
\keyword{internal}
