% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f.R
\name{zzo2f.anova}
\alias{zzo2f.anova}
\alias{zzo2f.aov}
\title{zzo2f method for ANOVA objects}
\usage{
\method{zzo2f}{anova}(x, digits = 3, ...)

\method{zzo2f}{aov}(x, digits = 3, ...)
}
\arguments{
\item{x}{An anova object.}

\item{digits}{Number of decimal places.}

\item{...}{Additional arguments passed to zzo2f.default.}
}
\value{
Invisibly returns the path to the output file.
}
\description{
zzo2f method for ANOVA objects

zzo2f method for aov objects
}
