% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f.R
\name{zzo2f.htest}
\alias{zzo2f.htest}
\title{zzo2f method for hypothesis tests}
\usage{
\method{zzo2f}{htest}(x, digits = 3, ...)
}
\arguments{
\item{x}{An htest object.}

\item{digits}{Number of decimal places.}

\item{...}{Additional arguments passed to zzo2f.default.}
}
\value{
Invisibly returns the path to the output file.
}
\description{
zzo2f method for hypothesis tests
}
