% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-formatting.R
\name{zzo2f_bold_col}
\alias{zzo2f_bold_col}
\title{Bold specific columns (Typst backend)}
\usage{
zzo2f_bold_col(cols)
}
\arguments{
\item{cols}{Integer or character vector.}
}
\value{
A \code{zzo2f_format} spec.
}
\description{
Bold specific columns (Typst backend)
}
\examples{
\dontrun{
zzo2f_bold_col(c("estimate", "p_value"))
}
}
