% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-inline.R
\name{zzo2f_engine}
\alias{zzo2f_engine}
\title{knitr engine for Typst-rendered tables}
\usage{
zzo2f_engine(options)
}
\arguments{
\item{options}{knitr chunk options.}
}
\value{
Engine output.
}
\description{
Used by \code{register_zzo2f_engine()} to handle
chunks of the form \code{```{zzo2f, zzo2f.caption="...", ...}}.
Chunk body should evaluate to a data frame.
}
\keyword{internal}
