% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-refs.R
\name{zzo2f_mark}
\alias{zzo2f_mark}
\title{Footnote marker for Typst cell content}
\usage{
zzo2f_mark(text, mark, type = c("symbol", "number", "alphabet"))
}
\arguments{
\item{text}{Character. Cell text to decorate.}

\item{mark}{Integer. Index into the selected marker family.}

\item{type}{One of \code{"symbol"}, \code{"number"},
\code{"alphabet"}.}
}
\value{
Character string with embedded Typst superscript.
}
\description{
Append a superscript footnote marker to a cell value
using Typst's \code{super()} function. Parallel to
\code{o2f_mark()} which emits \code{\\textsuperscript{}}.
}
\examples{
\dontrun{
df$value[1] <- zzo2f_mark("23.5", 1, "symbol")
}
}
