% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-preamble.R
\name{zzo2f_textlang}
\alias{zzo2f_textlang}
\title{Language directive for Typst (parallel to \code{babel()})}
\usage{
zzo2f_textlang(language)
}
\arguments{
\item{language}{Character. An ISO 639 language code (e.g.
\code{"en"}, \code{"de"}, \code{"fr"}).}
}
\value{
Character string \code{"#set text(lang: \"<code>\")"}.
}
\description{
Language directive for Typst (parallel to \code{babel()})
}
