% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-include.R
\name{translate_width}
\alias{translate_width}
\title{Translate a LaTeX width spec to a Typst width}
\usage{
translate_width(width)
}
\value{
Character. A Typst width expression.
}
\description{
Translate a LaTeX width spec to a Typst width
}
\keyword{internal}
