% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-include.R
\name{zzo2f-include}
\alias{zzo2f-include}
\title{Typst Figure-Inclusion Helpers}
\description{
Emit Typst source fragments that embed a rendered
table image (typically from a prior \code{zzo2f()} call) into a
surrounding \code{.typ} document. Parallel in purpose to the
\code{o2f_include*} family but using Typst primitives
(\code{#figure}, \code{#image}, \code{#grid}, \code{#place}).
}
\details{

}
