% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzo2f-preamble.R
\name{zzo2f_page}
\alias{zzo2f_page}
\title{Page directive for Typst (parallel to \code{geometry()})}
\usage{
zzo2f_page(
  margin = NULL,
  paper = NULL,
  landscape = FALSE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{margin}{Character or NULL. A Typst length
(\code{"2cm"}) or a named list like
\code{list(x = "2cm", y = "1.5cm")} for horizontal / vertical
margins, or \code{list(top = ..., bottom = ..., left = ...,
  right = ...)} for a full specification.}

\item{paper}{Character or NULL. Typst paper code (e.g.
\code{"a4"}, \code{"us-letter"}).}

\item{landscape}{Logical. Emit \code{flipped: true}.}

\item{width}{Character or NULL. Typst length, e.g. \code{"auto"}.}

\item{height}{Character or NULL.}
}
\value{
Character scalar Typst \code{#set page(...)} directive.
}
\description{
Page directive for Typst (parallel to \code{geometry()})
}
